refactor shop page components architecture
