+++ src/components/shop/BuyPageContent/index.tsx
@@ -5,8 +5,6 @@ import { Styled } from './styled'
 import { sortItems } from './types'
 import { Tabs } from '@components/common'
 import { BuyTabArticleList } from '@components/shop/ArticleList'
-import { ProfileBox } from '@components/shop/ProfileBox'
-import { myProfile } from '@mocks/fixture'
 import type { TradeBuyActivityType } from '@constants'
 import { TRADE_STATUS, TRADE_ACTIVITY_TYPE } from '@constants'
 import type { ArticlesElement } from '@types'
@@ -48,63 +46,58 @@ const BuyPageContent = (): ReactElement => {
 
   return (
     <div>
-      <Styled.ContentWrapper>
-        <ProfileBox {...myProfile} />
-        <Styled.Divider size="bold" />
-        <Styled.UserProductsWrapper>
-          <Tabs>
-            <Styled.SearchOptionsWrapper>
-              <Styled.TabsList>
-                {tradeBuyActivityList.map((tradeBuyActivity, index) => {
-                  const isCurrent = tabIndex === index
+      <Styled.Divider size="bold" />
+      <Styled.UserProductsWrapper>
+        <Tabs>
+          <Styled.SearchOptionsWrapper>
+            <Styled.TabsList>
+              {tradeBuyActivityList.map((tradeBuyActivity, index) => {
+                const isCurrent = tabIndex === index
 
-                  return (
-                    <Styled.Tab
-                      key={tradeBuyActivity[0]}
-                      onClick={handleTabClick}>
-                      <Styled.StatusButtonLabel>
-                        <Styled.Circle isCurrent={isCurrent} />
-                        <Styled.StatusButton
-                          isCurrent={isCurrent}
-                          type="button">
-                          <Styled.Text isCurrent={isCurrent}>
-                            {tradeBuyActivity[1]}
-                          </Styled.Text>
-                        </Styled.StatusButton>
-                        <Styled.Text color="grayScale50">1</Styled.Text>
-                      </Styled.StatusButtonLabel>
-                    </Styled.Tab>
-                  )
-                })}
-              </Styled.TabsList>
-              <SelectBox
-                colorType="none"
-                items={sortItems}
-                value="recently"
-                onChange={noop}
+                return (
+                  <Styled.Tab
+                    key={tradeBuyActivity[0]}
+                    onClick={handleTabClick}>
+                    <Styled.StatusButtonLabel>
+                      <Styled.Circle isCurrent={isCurrent} />
+                      <Styled.StatusButton isCurrent={isCurrent} type="button">
+                        <Styled.Text isCurrent={isCurrent}>
+                          {tradeBuyActivity[1]}
+                        </Styled.Text>
+                      </Styled.StatusButton>
+                      <Styled.Text color="grayScale50">1</Styled.Text>
+                    </Styled.StatusButtonLabel>
+                  </Styled.Tab>
+                )
+              })}
+            </Styled.TabsList>
+            <SelectBox
+              colorType="none"
+              items={sortItems}
+              value="recently"
+              onChange={noop}
+            />
+          </Styled.SearchOptionsWrapper>
+          <Styled.ProductListPanels>
+            <Tabs.Panel>
+              <BuyTabArticleList
+                activityType={
+                  tradeBuyActivityList[tabIndex][0] as TradeBuyActivityType
+                }
+                articles={articles}
               />
-            </Styled.SearchOptionsWrapper>
-            <Styled.ProductListPanels>
-              <Tabs.Panel>
-                <BuyTabArticleList
-                  activityType={
-                    tradeBuyActivityList[tabIndex][0] as TradeBuyActivityType
-                  }
-                  articles={articles}
-                />
-              </Tabs.Panel>
-              <Tabs.Panel>
-                <BuyTabArticleList
-                  activityType={
-                    tradeBuyActivityList[tabIndex][0] as TradeBuyActivityType
-                  }
-                  articles={articles}
-                />
-              </Tabs.Panel>
-            </Styled.ProductListPanels>
-          </Tabs>
-        </Styled.UserProductsWrapper>
-      </Styled.ContentWrapper>
+            </Tabs.Panel>
+            <Tabs.Panel>
+              <BuyTabArticleList
+                activityType={
+                  tradeBuyActivityList[tabIndex][0] as TradeBuyActivityType
+                }
+                articles={articles}
+              />
+            </Tabs.Panel>
+          </Styled.ProductListPanels>
+        </Tabs>
+      </Styled.UserProductsWrapper>
     </div>
   )
 }
