+++ src/components/shop/ReviewPageContent/index.tsx
@@ -45,50 +45,46 @@ const ReviewPageContent = (): ReactElement => {
 
   return (
     <div>
-      <Styled.ContentWrapper>
-        <ProfileBox {...myProfile} />
-        <Styled.Divider size="bold" />
-        <Styled.UserProductsWrapper>
-          <Tabs>
-            <Styled.SearchOptionsWrapper>
-              <Styled.TabsList>
-                {tradeReviewActivityList.map((tradeReviewActivity, index) => {
-                  const isCurrent = tabIndex === index
+      <ProfileBox {...myProfile} />
+      <Styled.Divider size="bold" />
+      <Styled.UserProductsWrapper>
+        <Tabs>
+          <Styled.SearchOptionsWrapper>
+            <Styled.TabsList>
+              {tradeReviewActivityList.map((tradeReviewActivity, index) => {
+                const isCurrent = tabIndex === index
 
-                  return (
-                    <Styled.Tab
-                      key={tradeReviewActivity[0]}
-                      onClick={handleTabClick}>
-                      <Styled.StatusButtonLabel>
-                        <Styled.Circle isCurrent={isCurrent} />
-                        <Styled.StatusButton
-                          isCurrent={isCurrent}
-                          type="button">
-                          <Styled.Text isCurrent={isCurrent}>
-                            {tradeReviewActivity[1]}
-                          </Styled.Text>
-                        </Styled.StatusButton>
-                        <Styled.Text color="grayScale50">1</Styled.Text>
-                      </Styled.StatusButtonLabel>
-                    </Styled.Tab>
-                  )
-                })}
-              </Styled.TabsList>
-            </Styled.SearchOptionsWrapper>
-            <Styled.ProductListPanels>
-              <Tabs.Panel>
-                <ReviewTabArticleList reviews={reviews} />
-              </Tabs.Panel>
-              <Tabs.Panel>
-                <ReviewTabArticleList reviews={reviews} />
-              </Tabs.Panel>
-              <Tabs.Panel>
-                <ReviewTabArticleList reviews={reviews} />
-              </Tabs.Panel>
-            </Styled.ProductListPanels>
-          </Tabs>
-        </Styled.UserProductsWrapper>
-      </Styled.ContentWrapper>
+                return (
+                  <Styled.Tab
+                    key={tradeReviewActivity[0]}
+                    onClick={handleTabClick}>
+                    <Styled.StatusButtonLabel>
+                      <Styled.Circle isCurrent={isCurrent} />
+                      <Styled.StatusButton isCurrent={isCurrent} type="button">
+                        <Styled.Text isCurrent={isCurrent}>
+                          {tradeReviewActivity[1]}
+                        </Styled.Text>
+                      </Styled.StatusButton>
+                      <Styled.Text color="grayScale50">1</Styled.Text>
+                    </Styled.StatusButtonLabel>
+                  </Styled.Tab>
+                )
+              })}
+            </Styled.TabsList>
+          </Styled.SearchOptionsWrapper>
+          <Styled.ProductListPanels>
+            <Tabs.Panel>
+              <ReviewTabArticleList reviews={reviews} />
+            </Tabs.Panel>
+            <Tabs.Panel>
+              <ReviewTabArticleList reviews={reviews} />
+            </Tabs.Panel>
+            <Tabs.Panel>
+              <ReviewTabArticleList reviews={reviews} />
+            </Tabs.Panel>
+          </Styled.ProductListPanels>
+        </Tabs>
+      </Styled.UserProductsWrapper>
     </div>
   )
 }
